gemma27b/src/requirements.txt
torch>=2.3
transformers>=4.43
accelerate>=0.33
huggingface-hub>=0.23
Pillow>=10.3
